Extract the NWS persistence transaction script #161
